"best epoch" now has negligible threshold
Protects against a subtle bug introduced when saving training history, 
where values were rounded to some fixed decimal point. To accomplish 
this, format strings are saved to a new attribute on initialization, 
fmt_dict. Then, every time the best is searched for, we format the 
metric as a string then back again.