Updates to TrainingStateController, including semantics bug fix (#28)
* Changed default epch of load_model_and_optimizer_for_epoch

* Added load_model_for_epoch, test

* Added "strict" flag to load_*_for_epoch in TrainingStateController

* "best epoch" now has negligible threshold

Protects against a subtle bug introduced when saving training history, 
where values were rounded to some fixed decimal point. To accomplish 
this, format strings are saved to a new attribute on initialization, 
fmt_dict. Then, every time the best is searched for, we format the 
metric as a string then back again.